refactoring: graphql config 위치 변경
* 일관성을 위해 별도 config 폴더로 이동
* add todos
