+++ backend/node/graphQL/config/config.js
@@ -1,3 +1,4 @@
+// todo .env 로 모드별 config 정하기
 const options = {
 	port: 8000,
 	endpoint: "/graphql",
